Initial commit: Design revamp and implementation of music player, caching
Revamped design: As midway, dev team found out that the design with the waveform is not plausible, we have decided to redesign the whole screen. Waveform.js has been removed from Main.js and revamped the layout of the screen

Implementation of Music player: this uses react-player from npm

Caching: The search results are now cached in the browser. Once the query used is used again, system will load the previously loaded results. The volume of the video is also cached

WARNING: This commit contains a mini-defect: The seek slider of the video is laggy: cause is found at ControlBar.js:72. The component lags as the component sets the seek value, and once the seek value changes, the seek value also sets the value of the seek slider. Kindly fix afterwards
Co-Authored-By: maruby <maruby@users.noreply.github.com>